f2_event_to_entities/requirements.txt
psycopg2-binary
phonenumbers
python-dotenv
kafka-python
